proptest: introduce property based testing
This patch introduces property based testing via the `proptest` crate.
In order to also benefit from these property based tests in wasm
environments a forked version of the `proptest` crate is used in order
to work around a platform issue with the upstream build.

Proptests for roundtripping Display/Fromstr impls for Address, Digest,
and Base64ArrayN are also added.