sui-rpc: add sui.rpc.v2 proto package
Update the vendored protos to include the new stable sui.rpc.v2 package.
For now all we are doing is providing the new proto package, in addition
to the old v2beta2 one, in order to allow for folks to migrate. After a
few release we'll be able to remove the older schema.